Private key insertion (#37)
## Overview
Implements functionality allowing users to pass an RSA-encrypted .pem private key into the Docker container at runtime, as a plaintext CLI argument, reconstructing the key file within the "~/.ssh" directory within the container. This approach avoids the user having to place the key in a file on the host machine but safety measures should be taken to ensure that the plaintext key is kept secure (see documentation). 

## Commits included
* Implements script to build private_key.pem
     * Implements a shell script which reads a raw private key from a docker
environment variable (if set) and rebuilds a private key into the
container at runtime.
* Updates dockerfile to run key builder
     * Updates dockerfile to run the private key builder and also adds execute
permissions to the key builder script.
* Removed "skipped private key" output
     * Removed output informing the user that the private key reassembly stage
had been skipped when the environment variable is not set, as this is
just ominous and overcomplicates the output.
* Adds key-insertion documentation
     * Adds documentation describing how to use the key-insertion functionality
and additional safety measures which should be taken alongside this.